+++ fatiando/inversion/regularization.py
@@ -112,7 +112,7 @@ class Damping(Objective):
     def __init__(self, nparams):
         super(Damping, self).__init__(nparams, islinear=True)
 
-    def hessian(self, p):
+    def _get_hessian(self, p):
         """
         Calculate the Hessian matrix.
 
@@ -130,7 +130,7 @@ class Damping(Objective):
         # This is cheap so there is no need to cache it
         return 2*scipy.sparse.identity(self.nparams).tocsr()
 
-    def gradient(self, p):
+    def _get_gradient(self, p):
         """
         Calculate the gradient vector.
 
@@ -151,7 +151,7 @@ class Damping(Objective):
             grad = 2.*p
         return grad
 
-    def value(self, p):
+    def _get_value(self, p):
         """
         Calculate the value of this function.
 
@@ -235,7 +235,7 @@ class Smoothness(Objective):
         self._cache['hessian'] = {'hash':'',
                                   'array':2*safe_dot(fdmat.T, fdmat)}
 
-    def hessian(self, p):
+    def _get_hessian(self, p):
         """
         Calculate the Hessian matrix.
 
@@ -252,7 +252,7 @@ class Smoothness(Objective):
         """
         return self._cache['hessian']['array']
 
-    def gradient(self, p):
+    def _get_gradient(self, p):
         """
         Calculate the gradient vector.
 
@@ -273,7 +273,7 @@ class Smoothness(Objective):
             grad = safe_dot(self.hessian(p), p)
         return grad
 
-    def value(self, p):
+    def _get_value(self, p):
         """
         Calculate the value of this function.
 
@@ -452,7 +452,7 @@ class TotalVariation(Objective):
         self.beta = beta
         self._fdmat = fdmat
 
-    def value(self, p):
+    def _get_value(self, p):
         """
         Calculate the value of this function.
 
@@ -469,7 +469,7 @@ class TotalVariation(Objective):
         """
         return numpy.linalg.norm(safe_dot(self._fdmat, p), 1)
 
-    def hessian(self, p):
+    def _get_hessian(self, p):
         """
         Calculate the Hessian matrix.
 
@@ -489,7 +489,7 @@ class TotalVariation(Objective):
                                       0).tocsr()
         return safe_dot(self._fdmat.T, q_matrix*self._fdmat)
 
-    def gradient(self, p):
+    def _get_gradient(self, p):
         """
         Calculate the gradient vector.